+++ client/shared/src/api/extension/test/extensionHost.hover.test.ts
@@ -66,7 +66,7 @@ describe('getHover from ExtensionHost API, it aims to have more e2e feel', () =>
             { isLoading: true, result: null },
             {
                 isLoading: false,
-                result: { contents: [textHover('a1').contents], alerts: [] },
+                result: { contents: [textHover('a1').contents], alerts: [], aggregatedBadges: [] },
             },
         ])
         results = []
@@ -79,13 +79,14 @@ describe('getHover from ExtensionHost API, it aims to have more e2e feel', () =>
         expect(results).toEqual<MaybeLoadingResult<HoverMerged | null>[]>([
             {
                 isLoading: true,
-                result: { contents: [textHover('a2').contents], alerts: [] },
+                result: { contents: [textHover('a2').contents], alerts: [], aggregatedBadges: [] },
             },
             {
                 isLoading: false,
                 result: {
                     contents: ['a2', 'b'].map(value => textHover(value).contents),
                     alerts: [],
+                    aggregatedBadges: [],
                 },
             },
         ])
@@ -98,7 +99,7 @@ describe('getHover from ExtensionHost API, it aims to have more e2e feel', () =>
             { isLoading: true, result: null },
             {
                 isLoading: false,
-                result: { contents: [textHover('a3').contents], alerts: [] },
+                result: { contents: [textHover('a3').contents], alerts: [], aggregatedBadges: [] },
             },
         ])
     })