+++ client/shared/src/api/integration-test/languageFeatures.test.ts
@@ -23,6 +23,7 @@ describe('LanguageFeatures (integration)', () => {
         labeledProviderResults: labels => ({
             contents: labels.map(label => ({ value: label, kind: MarkupKind.PlainText })),
             alerts: [],
+            aggregatedBadges: [],
         }),
         providerWithImplementation: run => ({ provideHover: run } as sourcegraph.HoverProvider),
         getResult: (uri, extensionHostAPI) =>