+++ internal/search/query/query.go
@@ -87,19 +87,25 @@ func InitStructural(in string) step {
 	return Init(in, SearchTypeStructural)
 }
 
+func Run(step step) ([]Node, error) {
+	return step(nil)
+}
+
 // Pipeline processes zero or more steps to produce a query. The first step must
 // be Init, otherwise this function is a no-op.
-func Pipeline(steps ...step) (Q, error) {
+func Pipeline(steps ...step) (Plan, error) {
 	nodes, err := sequence(steps...)(nil)
 	if err != nil {
 		return nil, err
 	}
 
+	plan := make([]Q, 0, len(nodes))
 	for _, disjunct := range Dnf(nodes) {
 		err = validate(disjunct)
 		if err != nil {
 			return nil, err
 		}
+		plan = append(plan, Q(disjunct))
 	}
-	return nodes, nil
+	return plan, nil
 }